Skip page reload checks in older Webkits.
They don't suppor the history API, and it looks like the backporters who
put them in Lucid for us aren't doing so any more, so it looks like we
can't sensibly get hold of a browser that passes these tests.

This is an interim fix until we either get hold of such a browser in a
reliable (i.e. Puppet-y) way or remove the tests entirely.