Set cache expiry to 30 minutes
The beauty of smart answers is that all of the information you place
into them is then added into the URL. As a result, the generated page
views for ANY of the steps can be cached.

However, we're setting all cache headers to a baseline of 30 mins to
start off with and then we can tweak each format as/when we need to.