Add logic for starting and stopping in same year
Adds an additional question (of type "optional
date") asked for each 'starting' child, which
will be saved as a date or "no".

We then adjust the ranges we provide to the
calculator to use this early leaving date instead
of the end of the tax year, if available.

Additional fix to avoid showing the "children
leaving" question if there aren't existing
children, as in such case, the answer must be
zero anyway.