Validate income inside the flow, not the calculator
This buys us two things:-
- It's easier to test the validation in the smart 
answer.
- If you happen to have greater pension contributions
and gift aided donations than your income, you will
have an adjusted net income of < 1, so you should
get the maximum amount. 

The existing validation is still in the calculator
to maintain backwards compatibility with V1, but
it is bypassed by a key in the hash parameter to
the initialize method.