Use explicitly separate nodes for employee/employer
The intention here is that we'll need separate copy. Rather than
hacking this with PhraseList objects it's far easier to read by
doubling the number of nodes in this smart answer.

In this case we're making 3 nodes look like 6.

I checked this over with Steve in a discussion we had.