Fix the CountrySelect initialize to not load options
This was unnecessary and seemed to be incorrect, maybe a hangover from
inheriting from MultipleChoice.  This change means it doesn't need to
load the countries from the API on initialise, which means it doesn't
need to load them when registering with Panopticon... which is nice.

It no longer inherits form MultipleChoice as there was no real reason to.