Look up the asset host using Plek
Fall back to asking Plek for the asset host if the STATIC_DEV 
environment variable isn't set. This only applies in development and 
will result in the same behaviour as present, however it makes it easier 
to override the URL for a particular service across the whole stack 
using environment variables.