Don't blow up when Content API is unavailable
In Smart Answers, the Content API is only used to populate the related 
items, sections, and associated metadata. If an artefact doesn't exist, the app continues to operate without the metadata.

However, if the Content API is completely unavailable (eg. if you are 
running Smart Answers in isolation), Smart Answers blows up with a
GdsApi::HTTPErrorResponse.

This change catches that exception and returns an empty hash, so that 
the page can still be rendered without the artefact metadata.