Use next_node as fallback in multiple choice question
Instead of the next_node command overriding the other calculations, use
it as a fallback if no other match is found.

Only one flow (the minimum_wage.rb shared logic) relied on the original
overriding behaviour. I've updated that to work with the new semantics
of next_node being a fallback.

I have verified that no other flows rely upon this behaviour (by writing
a script which parses the flow source code using the ruby parser).