Provide indexable_content when registering with Panopticon
For now it only uses text from the cover sheet, as i'm not convinced node text (like
title / body) is useful for indexing. After chatting with Richard Boulton he agreed.

I've tried to code this in a way that would be easy to extend to pull content from
other nodes, questions and outcomes, it should be as simple as:
 - @flow.nodes.map { |node| node_text(node) }.join