Make jenkins test merge result of branches
As we do in manyother repos, eg;
- https://github.com/alphagov/rummager/commit/698c2c896f9aa227dbd7dc8600c043008f084db7

'Almost all of the time we are writing code to be merged into master
sooner rather than later. Rather than testing the HEAD of the branch in
Jenkins this commit adds a command to the build script which merges the
branch with master before building it.'

This will help us catch some test failures on master that
had passed on branch builds. Catch errors earlier.