Remove all references to 'undo' links
The original commit that added tracking to these links appears to have gone in in April as a temporary measure to support some design reckons. This commit proposes to take out the link tracking and css references to 'undo', as we no longer seem to have the concept of an 'undo' link in the markup.

Original ref from `git log -p -S 'This is short lived tracking'`:

```git
commit cfefadf12a3844bb03838b9b2d9cd98fe2239bb6
Author: David Singleton <david.singleton@digital.cabinet-office.gov.uk>
Date:   Mon Apr 28 13:38:39 2014 +0000
```