Remove usage of `Date.today` inside the constant
This usage of Date.today caused a bug in production earlier. When the
application was started the year was 2014, and the end date that
was stored in the constant (and was never updated) was "2014-12-31".

I don't think that this is the ultimate solution to this issue, but
I'm committing this change as an indication that using Date.today
inside a constant means that the value won't be changed unless
the application is restarted.