Make landing page two columns
So that the landing page looks like other start pages on GOV.UK wrap the
main in a `grid-row` element.

Also most the related-items div into the show template and toggle it on
`presenter.started?`. This is the same logic that was toggling it before
but before it was being inserted using a yield. The grid layout requires
it to be in a slightly different place and just using an if seems to
make sense.