Remove workplace-pension Smart Answer flow
This has been archived (https://www.pivotaltracker.com/story/show/87305558)
but was still registering with the panopticon during deploy, as the flow
had not been removed.

This removes the flow entirely, rather than setting it to draft, as for
smart-answers pulling something out of archive is equivilent to pulling
it out of git history, as git is used to version/archive content.