Ignore the extra day for leap years when displaying state_pension_age
The Calculator correctly shows that these people reach SPA on 1 March if their Spa falls on their birthday in a year that isn’t a leap year. However, the calculation of what their SP age is shows that it is XX years and 1 day. This is potentially unhelpful as the calculator would also show that people born a day before them and a day after them would have a State Pension age of XX years, thereby creating a situation where people born on 29 February may feel they are being penalised compared to people born on 1 March.