Fix raw yaml being rendered on the outcome page
The bug applies to all countries without incomplete application deadline.
e.g. /overseas-passports-v2/y/belarus/renewing_new/adult
The issue was that for those countries a new phrase key '' was added.
Later it was interpolated with the rest of the translation scope thus
resulting in the whole translation hash in that scope to be rendered
on the page.