'What you need to do' amendments for OS marriage in Jordan
TODO: add the link to the document template

As per ticket:
CHANGE RELATES TO: all opposite sex marriage outcomes in Jordan

1)
Section: What you need to do
ADD to the start of the section:

You need to have been living in the country where you intend to marry for 21 days.
If you’re getting married at the Sharia Court, you’ll need to provide a
[single declaration document](see zendesk for attachment) to prove you’re allowed to marry.
Make an appointment at the local British embassy or consulate to swear an affidavit
(written statement of facts) that you’re free to marry.
You’ll need to bring the single declaration, your passport, your partner’s passport and pay a fee.
---
2)
CHANGE:
You may be asked to provide a certificate of no impediment (CNI) or a similar
document to prove you’re allowed to marry.
TO:
For non-Islamic marriages, you may be asked to provide a certificate of
no impediment (CNI) or a similar document to prove you’re allowed to marry.