Remove Affidavit attachments from some CNI countries
Fix to previous incomplete logic.

As per ticket:
This change relates to British national marrying in a CNI country but is resident in the UK.

CNI Countries affected:

Albania, Algeria, Angola, Armenia, Austria, Azerbaijan, Bahrain, Bolivia, Bosnia Herzegovina, Bulgaria, Croatia, Cuba, Estonia, Georgia, Greece, Iceland, Italy, Japan, Kazakhstan, Kuwait, Kyrgyzstan, Libya, Lithuania, Luxembourg, Macedonia, Mexico, Moldova, Montenegro, Nicaragua, Norway, Poland, Russia, Spain, Serbia, Sweden, Tajikistan, Tunisia, Turkmenistan, Ukraine, Uzbekistan and Venezuela.

The advice is correct about obtaining a CNI from a UK Registrar and the possibility that it may need to be legalised, but there is additional text and attachments towards the end that are not relevant; namely the Notice of Marriage and Affidavit downloads which should be taken out when the tool is for British national resident in UK marrying in a CNI country. Grateful if the following text can be taken out from the relevant outcome pages:

BEGINS

You’ll need to fill in forms for your notice of marriage, and an oath or affidavit (written statement of facts) stating that you’re free to marry.

You can download and fill in (but not sign) the forms in advance.

Download ‘Notice of marriage’ (PDF, 96KB)

Download ‘Affidavit for marriage’ (PDF, 58KB)

ENDS