Extract rates for married couples allowance
This is a refactoring that should make it easy to make the changes
in future. Thos changes happen yearly and every time a new developer
has to understand what's going on. All that will need to change now
is going to be the yaml file, unless there are changes to the logic.