Fix a randomly failing test
To reproduce the issue, run `rake test TESTOPTS="--seed=25874"`
or `rake test TESTOPTS="--seed=45887"`.
The issue occurs due to the fact that presenter_test.rb loads
smardown plugins before they are tested, the result is cached
and when the plugins are tested in consecutive tests, they are
not actually loaded in the cached values and fail.

Make sure plugin fixtures cache is purged and correct fixtures
are loaded before testing example-plugins