Introduce outcome_1_scotland
Adding the scenarios highlighted a problem with outcome_1 when reached via the
Scotland path. The 'outcome_1' outcome uses the value of `estate_over_250000`
to decide whether to display the link to information about Inheritance Tax.
This variable isn't available when the outcome is arrived via the Scotland path
and I'm not aware of any way to check that a variable exists before checking
its value. Introducing the 'outcome_1_scotland' outcome seems like the best fix
for the moment.