Allow Logstasher to be disabled
We can't deploy to Heroku if Logstasher is enabled. We see errors similar to:

> Errno::ENOENT: No such file or directory @ rb_sysopen -
> /tmp/build_0bd997e043bd4565bfc5623757dc05ac/log/production.json.log

We think this is because the app is created in the /tmp directory and is then
moved to /app when Heroku has finished doing its thing.

We don't expect the `RUNNING_ON_HEROKU` environment variable to be set on
preview or production so this change shouldn't have any affect on them.