Consistently use Date vs Time objects in integration test.
The test was responding to date questions with instances of `Time` in some
cases, but instances of `Date` in others. In reality a user can only respond
with a date to a date question, so it feels better to be consistent.

I think there might be clearer ways to build these date objects e.g.
`Date.today - 1.month - 21.days`, but I thought it was clearer to make this
change separately.