Add date_question option to parse response into Date object
Having parsed the response into a Date object it makes no sense to me to then
convert it back into a String. There seems to be a lot of code in the Ruby DSL
flavoured smart answers (and their calculators) which has to re-parse the date
strings. Also some date "variables" are Date objects and some are date strings
in these smart answers which is really confusing.

The plan is ultimately to always supply the response to a
`SmartAnswer::Question::Date` as a Date object and to only use Date objects
within the Ruby DSL and associated calculators. However, to get there
incrementally I'm introducing this option so we can convert one question or
smart answer at a time.

We think that it will be worth adopting this pattern for all question types,
but we have to start somewhere.