Use Date objects in appeal-a-benefits-decision
I've made the conversion relatively simplistic to make the change as clear as
possible. There are some obvious further simplifications that could be made,
but I'm going to leave those for now.

The outcomes for this smart answer do not display any dates, so we don't need
to worry about how they might be rendered.