Ensure Date objects are rendered using long format in locale file
Given that some date questions are now supplying their response as a `Date`
object rather than as a `String`, we need to make sure they are being rendered
correctly. This is mostly already taken care of by the `Date` `when` clause in
`NodePresenter#value_for_interpolation`. However, we do need to make a change
to `DateQuestionPresenter#response_label` so that it can handle `Date` objects
as well as date strings.