Use Date objects in calculate-state-pension
The associated `StatePensionAmountCalculator`, its unit test and the
integration test for the smart answer have all also been updated to work with
`Date` objects.

I've made the conversion purposely simplistic to make the change as clear as
possible. There may be some further simplifications that could be made, but I'm
going to leave those for now.

The outcomes for this smart answer do display dates, but I've fixed the
`DateQuestionPresenter` to work with date questions which supply their response
as a `Date` object in an earlier commit.