Use Date objects in calculate-your-holiday-entitlement
The associated `Calculators::HolidayEntitlement`, its unit test and the
integration test for the smart answer have all also been updated to work with
`Date` objects.

I've made the conversion purposely simplistic to make the change as clear as
possible. There may be some further simplifications that could be made, but I'm
going to leave those for now.