Notify about missing phrases via Airbrake
in 3e9c2909758c480bfb4fd2897f20748d051739e6 a change was introduced
to silently fall back to a phrase key if the phrase is not defined
in the YML file. I think developers should know about cases like
this. Hence I'm adding airbrake notification, but leaving the UX
as it was at least initially. Perhaps later we can just raise the
exception so we know about it in dev/test environments as well as
staging/production.