Log missing phrase keys
in 3e9c290 a change was introduced
to silently fall back to a phrase key if the phrase is not defined
in the YML file. I think developers should know about cases like
this. Hence I'm logging it, but leaving the UX as it was at least
initially. Perhaps later we can just raise the exception so we know
about it in dev/test environments as well as staging/production.

This overwrites the changes in
https://github.com/alphagov/smart-answers/pull/1614 because they
generated too much noise in staging and messed up metrics and alerts