Add test for SmartdownAdapter::DateQuestionPresenter#to_response
Given the changes I've been making to date questions in Ruby/DSL-style Smart
Answer flows, I was a bit nervous that this method wasn't covered by any tests
and that its implementation might now be wrong.

It turns out that this method is only used when you choose to **change** the
answer to a previous date question. It's used to pre-select the relevant
options in the date question select elements. It turns out that the `input`
value passed into the `to_response` is always a `String` and so the
implementation is correct.

The `ChangingAnswerTest` integration test tests this behaviour for
Ruby/DSL-style Smart Answer flows, but the behaviour is untested for Smartdown
flows. It would be quite a bit of effort to setup an equivalent integration
test, so for now, I've gone with this unit test.

I spent quite a bit of time tracing the call stacks for both flow-types and I
thought it might help someone else if I include them here:

* Smartdown Flow

    show.html.erb
      _content.html.erb
        _current_question.html.erb
          SmartdownAdapter::QuestionPresenter#partial_template_name
            _date_question.html.erb
              SmartdownAdapter::DateQuestionPresenter#defaulted_day?
              SmartdownAdapter::DateQuestionPresenter#defaulted_month?
              SmartdownAdapter::DateQuestionPresenter#defaulted_year?
              CurrentQuestionHelper#default_for_date
              CurrentQuestionHelper#prefill_value_for (if previous_response param set)
                SmartdownAdapter::DateQuestionPresenter#to_response
              SmartdownAdapter::DateQuestionPresenter#default
              SmartdownAdapter::DateQuestionPresenter#start_date
              SmartdownAdapter::DateQuestionPresenter#end_date

* Smart Answer flow

    show.html.erb
      _content.html.erb
        _current_question.html.erb
          ::QuestionPresenter#partial_template_name
            _date_question.html.erb
              ::DateQuestionPresenter (undefined)
                ::QuestionPresenter (undefined)
                  ::NodePresenter#method_missing
                    SmartAnswer::Question::Date#defaulted_day?
                    SmartAnswer::Question::Date#defaulted_month?
                    SmartAnswer::Question::Date#defaulted_year?
              CurrentQuestionHelper#default_for_date
              CurrentQuestionHelper#prefill_value_for (if previous_response param set)
                ::DateQuestionPresenter#to_response
                  ::QuestionPresenter (undefined)
                    ::NodePresenter#to_response
                      SmartAnswer::Question::Date#to_response
              ::DateQuestionPresenter (undefined)
                ::QuestionPresenter (undefined)
                  ::NodePresenter#method_missing
                    SmartAnswer::Question::Date#default
              ::DateQuestionPresenter#start_date
              ::DateQuestionPresenter#end_date