Simplify logic in QuestionPresenter#partial_template_name
This logic was introduced in [1] amongst a bunch of unrelated changes and
there's no mention of these changes in the commit note. Also there are no unit
tests around the behaviour revealing the motivation behind it.

My assumption is that the intention was to identify the most specific partial
template for the node, based on its class or the classes of its ancestors.
However, I suspect there is a typo in the following line:

    template_names.first do |template_name|

I think `first` was meant to be `find`, but with the implementation as it
stands the block is never executed and the first value in the array of template
names is returned i.e. the template with its name based on the class of the
presenter itself and not any of its ancestors.

Since the intended behaviour has **never** been working, rather than fix it to
work, I've elected to simplify the code while retaining the current behaviour.
I've also added some basic unit test coverage for the method.

[1]: 1ac12b6b0be70aa5a816e707e240cbfdf67d5692