Removed unused DOB question partial template
This template was introduced in [1] ostensibly to support the
`can-i-get-a-british-passport` flow introduced in the same commit. However,
it was never actually used - the `date_of_birth` question [2] used a standard
`date_question` and thus the `_date_question.html.erb` partial template.

In any case this flow was removed entirely in [3], so I think it's safe to
remove it.

[1]: ae39fa485477592718df060b49f054c691788c41
[2]: https://github.com/alphagov/smart-answers/commit/ae39fa485477592718df060b49f054c691788c41#diff-c20178712cc1c62a55788dc3a3402944R31
[3]: 7ad7f10b4a88eecb71a002ed371b0e29f6837b5f