Add value_question option to parse response into Integer object
There seems to be a lot of code in the Ruby DSL flows which has to parse the
response from a `value_question` into an `Integer`. It feels as if it would be
simpler if the Ruby DSL flows always dealt with richer non-string value types
and the parsing was dealt with in one place in
`SmartAnswer::Question::Value#parse_input`.

This change means I'll be able to incrementally convert one question or smart
answer flow at a time to use the new option.