Use Integer objects in calculate-agricultural-holiday-entitlement
I've made the conversion relatively simplistic to make the change as clear as
possible. There are some obvious further simplifications that could be made,
but I'm going to leave those for now.

Neither of the new `Integer` response objects are rendered directly in outcomes
and in any case, since `12.to_s # => "12"`, I think they'd be rendered
correctly even if they were.