Fix I18n::MissingInterpolationArgument error
We started seeing this exception [1] in staging after an attempted deployment:

    I18n::MissingInterpolationArgument: missing interpolation argument :payday_offset_formatted in "What was the last normal payday before %{payday_offset_formatted}"

It can be reproduced using the following path:

    /maternity-paternity-calculator/y/paternity/no/2015-03-18/2015-03-31/yes/yes/yes/yes/yes/2015-04-01/two_weeks/2014-11-30

@tadast narrowed the problem down to this commit [2]. It seems as if I forgot
to add this `precalculate` block for the formatted version of the
`payday_offset` date. Unfortunately mistakes like this are not caught by the
integration tests, because they do not render the outcomes.

[1]: https://errbit.production.alphagov.co.uk/apps/533c35ae0da1159384044f5f/problems/5552047665786353be0a6300
[2]: 0b46e26501e1b21636974ec6d5328a8893c69a9e