Remove 'Already getting maternity allowance' context
It was incorrect: nested examples covered the cases where a user
is not getting maternity allowance

@floehopper says:

I've just cherry-picked this commit from @BenJanecke's
`statutory-sick-pay-updates` branch i.e. #1634.

In doing so I've had to resolve a bunch of conflicts in
`test/integration/smart_answer_flows/calculate_statutory_sick_pay_test.rb`
which were all due to a combination of the change in nesting of the contexts in the
original commit [1] versus the addition of some new contexts in [2].

[1]: 27d7b96818433a1bb56abfe528eb0efb9e309de8
[2]: f3eea3b8bf1708294afaa40cabda19e8902b385d