Provide a less strict integer parsing option for value_questions
Some `value_questions` are using `String#to_i` in `calculate` blocks to parse
string responses into integers. Ideally we'd be consistently using the same
parsing mechanism for all value questions expecting an integer response, but
it's difficult to determine whether the differences are by accident or design.
So for the moment, providing two mechanisms i.e. using `Integer()` or
`String#to_i` feels like a small improvement and at least makes the difference
more explicit. We can decide whether to try to combine these alternative
mechanisms later.

It's worth noting that using the new `:to_i` parse option instead of converting
the string response within a `calculate` block in the flow will result in a
few small changes in behaviour. If the user enters a non-integer string
response, the **parsed** `Integer` will be used in (a) the URL path segment;
(b) the "Previous Answers" section; and (c) the pre-filled value if the user
clicks on the "Change" link for the relevant previous question.

Before:

* User responds with "1.1"
* URL contains "/1.1"
* "Previous Answers" contains "1.1"
* Text input contains "1.1" after clicking "Change" within "Previous Answers"

After:

* User responds with "1.1"
* URL contains "/1"
* "Previous Answers" contains "1"
* Text input contains "1" after clicking "Change" within "Previous Answers"

I think this new behaviour, while different, is actually better. Also the old
URL paths and query strings will still work sensibly. So I'm happy to go with
it.