Use Integer objects in maternity-paternity-calculator
I've made the conversion relatively simplistic to make the change as clear as
possible. There are some obvious further simplifications that could be made,
but I'm going to leave those for now.

I'm using the newly introduced `parse: :to_i` option to convert the string
responses into integers which mostly matches the existing behaviour, but it's
worth noting that there are a few small differences in behaviour which are
explained in the commit note for the commit [1] in which the option was
introduced.

None of the new `Integer` response objects are rendered directly in outcomes.

[1]: "Provide a less strict integer parsing option for value_questions"