Use Integer objects in calculate-your-child-maintenance
Although the original implementation was using `String#to_i` to convert the
response into a `String`, the "validation" in the `calculate` block meant that
the behaviour visible to the user was the same as using the `parse: Integer`
option for the `value_question`.

For example, the following all used to raise a `SmartAnswer::InvalidResponse`
exception which caused the user to see the default error message: "Please enter
a number".

* `"".to_i.to_s    # => "0" and "0" != ""`
* `"x".to_i.to_s   # => "0" and "0" != "x"`
* `"1.5".to_i.to_s # => "1" and "1" != "1.5"`

Similarly, the following will all raise `ArgumentError` in
`SmartAnswer::Question::Value#parse_input` which will be rescued in
`SmartAnswer::Flow#process` causing the user to see the same default error
message as before:

* `Integer("")     # => ArgumentError`
* `Integer("x")    # => ArgumentError`
* `Integer("1.5")  # => ArgumentError`

The only fly in the ointment is when the raw response is `nil`, because:

* `nil.to_i.to_s   # => "0" and "0" != nil`
* `Integer(nil)    # => TypeError`

And `TypeError` is not rescued in the same way as `ArgumentError`. However, it
doesn't seem to be possible to set the raw response to `nil` via the user
interface. Based on the somewhat obscure comment I guessed that maybe some
tests were relying on being able to set the raw response to `nil`, but this
doesn't appear to be the case. If this does turn out to be a problem, we can
rescue the `TypeError` in `SmartAnswer::Question::Value#parse_input` and raise
an appropriate `SmartAnswer::InvalidResponse` exception instead.

The new `Integer` response object is not rendered directly in any outcomes,
so we don't need to worry about anything changing in that respect.