Rely on bundler to load syck gem
The syck gem is included in the Gemfile and so we don't need the
`require 'rubygems'` statement, as long as we move the `require 'syck'`
statement below the `require 'bundler/setup'` statement. `yaml` is part of
the standard library.

This means that our application boot is more like that of a standard Rails app
and allows people to avoid needing to use `bundle exec` to run the tests using
this technique [1].

[1]: https://tomafro.net/2012/06/tip-bundler-with-binstubs