Add St Martin and Saint-Barthélemy to marriage abroad logic
Those two countries were recently added to the worldwide API
but had no corresponding logic to handle them. This change leads
to "no British consular facilities" outcome for both countries,
for both same and opposite sex marriages, as advised by FCO.

Those two countries also don't have many embassies, if any,
so we're always advising users to contact LOCAL relevant authorities
instead of embassies.

Note that currently the outcome has an 'os' (opposite sex) reference
in it and all the phrase keys. This is incorrect as of this commit,
but I have not renamed it here to avoid merge conflicts, as there are
changes in another branch that already started renaming those keys.