Require simplecov gems after bundler setup.
The `simplecov` & `simplecov-rcov` gems are included in the `Gemfile` and so if
we were just relying on the bundled gems, the code wasn't working when the
`TEST_COVERAGE` environment variable is set. I assume that these gems are
installed as system gems on the CI machine.

Moving the `require` statements after loading the Rails environment ensures
that the gems are loaded from the bundle.

This change does potentially mean that we lose coverage for files like
`config/environment.rb`, `config/application.rb` & `config/boot.rb`, but
these don't appear in the existing CI coverage reports [1] and so I suspect
they may be excluded by default.

[1]: https://ci-new.alphagov.co.uk/job/govuk_smart_answers/2396/rcov/