Add script to generate input responses for Smart Answers
This script creates a YAML file containing questions and responses for a given
flow.

It iterates over the questions in a flow and finds the possible responses for
those questions that list them (checkbox questions, radio button questions and
country lists).

For the questions that don't have a list of responses, we emit a TODO item in
the YAML file explaining that you need to manually add some responses.

We can use these completed set of responses to exercise the flows and
automatically generate integration-like tests for the existing Smart Answers.