Merge two phrases for marriage abroad
Replace occurences of consular_cni_os_local_resident_not_germany_or_spain_or_foreign_resident_not_germany
with consular_cni_os_download_documents_notary_public as the text
is almost the same. The former mentions an embassy and the latter
does not, therefore the latter is more accurate, because everything
has to be done in either the ceremony country or the UK now.