Add script to generate expected outcomes for Smart Answers
This script generates a YAML file containing the the expected outcomes of
exercising the flow with various sets of input responses.

The input responses are read from the YAML files generated by the
'generate-questions-and-responses-for-smart-answer.rb' script.

We can use this YAML file to automatically generate integration tests for
existing Smart Answers.