Add regression tests for Smart Answers
Request each set of input responses (from the responses-and-expected-results
YAML files) and save the output. The plan is to store the output files in git
so that we can diff the output and fail the test if anything has changed. If
the test fails but the output has changed by design (i.e. because the flow has
legitimately changed) then we should commit the revised output files to the
repo.

We're hoping that having these sort of tests will increase our confidence to
implement some of the larger changes we're hoping to make, e.g. the switch to
ERB templates.

In the short term, we imagine these regression tests slowly replacing the
existing integration tests. In the longer term, we hope to be able to get rid
of/reduce the number of these regression tests by having clearly separated,
unit-tested code.

Note the use of `Timecop` to freeze time in the setup of these tests. I'm doing
this to avoid the HTML output changing each day because of changes in the "Last
updated" date printed at the bottom of each Smart Answer[1]. In production this
should return the time the REVISION file was changed but in development it
returns the current date[2].

I initially tried an alternative approach: Replacing the text of the
`.modified_date` element to something known so that it didn't cause the tests
to fail. @floehopper convinced me that stubbing time was probably a better
approach in this case.

[1]: https://github.com/alphagov/smart-answers/blob/master/app/views/smart_answers/_smartanswer_metadata.html.erb#L4
[2]: https://github.com/alphagov/smart-answers/blob/master/app/helpers/application_helper.rb#L3