Update questions-and-responses for student-finance-calculator
* Added response for `how_much_are_your_tuition_fees_per_year?`

* Added responses for `whats_your_household_income?`

* Combined responses for
`do_any_of_the_following_apply_uk_full_time_students_only?`

* Combined responses for `do_any_of_the_following_apply_all_uk_students?`

I've added responses that aim to exercise all of the code in the flow. I used
@floehopper's work that allows us to use Simplecov to see Flow code coverage
(currently in a branch) to come up with these responses.

I've combined responses where I didn't need to test the multiple checkbox
values independently of each other. This reduces the number of combinations of
responses that we use to run tests, which ultimately reduces the number of
tests.