Avoid order-dependent test failure in TranslatorLinks
I was seeing a bunch of tests fail in the following two integration tests:

* `test/integration/smart_answer_flows/register_a_birth_test.rb`
* `test/integration/smart_answer_flows/register_a_death_test.rb`

I tracked the problem down to the fact that the following unit test was running
**before** these two integration tests and the stubbed version of the `Hash`
was being memo-ised in a class instance variable within the following method:

`SmartAnswer::Calculators::TranslatorLinks.translator_links`

I was able to reproduce the failures using the following command:

    bundle exec ruby -I"lib:test" -I".bundle/gems/gems/rake-10.4.2/lib" ".bundle/gems/gems/rake-10.4.2/lib/rake/rake_test_loader.rb" "test/unit/calculators/translator_links_test.rb" "test/integration/smart_answer_flows/register_a_birth_test.rb" --seed=27340 --verbose

My fix is to stop memo-ising the `Hash`. I don't think this will lead to any
significant performance problems, because the YAML file isn't very big and in
any case, it will only be loaded once per __instance__ of the `TranslatorLinks`
class.