Remove minitest-reporters Gem
This was added in bda81db4d948d79ba43a38e740c6ea657ece8be3.

It's configured to display the test output in a different format in Jenkins.

I find the test output in Jenkins hard to read and it makes it hard to see any
errors and failures.

I made a number of tests fail to demonstrate the difference in output between
using the SpecReporter and the standard Minitest output. You can see the
difference in my gist[1].

[1]: https://gist.github.com/chrisroos/44696c502f2dac8d6270