Update OutcomePresenter to support ERB templates
This commit doesn't do very much in isolation. It's setting us up to be able to
use ERB templates for outcomes, rather than a combination of `PhraseList`s and
the `NodePresenter`.

We're expecting ERB templates to be named after the outcome, and to live in
lib/smart_answer_flows/flow-name. For example; Given a flow named
'example-flow' and an outcome named 'outcome-1', then we'd expect the ERB
template to be available at
'lib/smart_answer_flows/example-flow/outcome-1.txt.erb'.